[2365] Moved user_come_from_backlink method into app controller
