+++ app/controllers/trainees/lead_schools_controller.rb
@@ -59,7 +59,7 @@ module Trainees
     end
 
     def step_wizard
-      @step_wizard ||= Wizards::Schools::StepWizard.new(trainee: trainee, page_tracker: page_tracker)
+      @step_wizard ||= Wizards::SchoolsStepWizard.new(trainee: trainee, page_tracker: page_tracker)
     end
 
     def validate_form_completeness
@@ -67,10 +67,6 @@ module Trainees
       return if user_came_from_backlink?
 
       redirect_to step_wizard.start_point if step_wizard.start_point.present?
-    end
-
-    def user_came_from_backlink?
-      session[:origin_path]&.include?("edit")
     end
   end
 end
