+++ spec/lib/wizards/schools_step_wizard_spec.rb
@@ -3,33 +3,31 @@
 require "rails_helper"
 
 module Wizards
-  module Schools
-    describe StepWizard do
-      subject { described_class.new(trainee: trainee) }
+  describe SchoolsStepWizard do
+    subject { described_class.new(trainee: trainee) }
 
-      describe "#start_point" do
-        context "when theres no lead or employing school" do
-          let(:trainee) { create(:trainee, :school_direct_salaried) }
+    describe "#start_point" do
+      context "when theres no lead or employing school" do
+        let(:trainee) { create(:trainee, :school_direct_salaried) }
 
-          it "returns the lead school step" do
-            expect(subject.start_point).to eq "/trainees/#{trainee.slug}/lead-schools/edit"
-          end
+        it "returns the lead school step" do
+          expect(subject.start_point).to eq "/trainees/#{trainee.slug}/lead-schools/edit"
         end
+      end
 
-        context "when theres a lead school but no employing school" do
-          let(:trainee) { create(:trainee, :school_direct_salaried, :with_lead_school) }
+      context "when theres a lead school but no employing school" do
+        let(:trainee) { create(:trainee, :school_direct_salaried, :with_lead_school) }
 
-          it "returns the employing school step" do
-            expect(subject.start_point).to eq "/trainees/#{trainee.slug}/employing-schools/edit"
-          end
+        it "returns the employing school step" do
+          expect(subject.start_point).to eq "/trainees/#{trainee.slug}/employing-schools/edit"
         end
+      end
 
-        context "when all the forms are complete" do
-          let(:trainee) { create(:trainee, :school_direct_salaried, :with_lead_school, :with_employing_school) }
+      context "when all the forms are complete" do
+        let(:trainee) { create(:trainee, :school_direct_salaried, :with_lead_school, :with_employing_school) }
 
-          it "returns nil" do
-            expect(subject.start_point).to eq nil
-          end
+        it "returns nil" do
+          expect(subject.start_point).to eq nil
         end
       end
     end
